Move get_wider_type to type_coercion module
